edited

+++ program2/array_avg/src/App.java
@@ -1,7 +1,7 @@
 public class App {
     public static void main(String[] args) {
-        int values[] = { 1, 2, 3, 4, 5 };
-        int result = 0;
+        double values[] = { 1, 2, 3, 4, 5 };
+        double result = 0;
 
         for (int i = 0; i < 5; i++) {
             result += values[i];
